refactor: extract buildModel to reusable server helper and add INSIGHTS cache tag
